Renamed fonction to functions
